Extract the server endpoints in preparation for future refactorings
